Enhance MAUI UN Trace Tool
In an apparent oversight during the initial implementation, UI for choosing terminal and fraction along settings for starting points was omitted. This commit adds a slider and picker for those two settings (they only appear for some starting points) and also enhances the appearance of the starting point delete button.